Cleanup of build process.
- cleanups to avoid multiple definitions: no more forcing linker to
  allow multiple definition
- redifinition of doubleEq removed
- some type-conversions made explicit to avoid warnings
- set compiler to use c/c++-17
- separate settings for MSVC and MS Windows: llvm-clang can also work

+++ basecode/Conv.h
@@ -566,7 +566,7 @@ template< class T > class Conv< vector< vector< T > > >
         {
             static vector< vector< T > > ret;
             ret.clear();
-            unsigned int numEntries = **buf; // first entry is vec size
+            unsigned int numEntries = (unsigned int)**buf; // first entry is vec size
             ret.resize( numEntries );
             (*buf)++;
             for ( unsigned int i = 0; i < numEntries; ++i ) {
@@ -641,7 +641,7 @@ template< class T > class Conv< vector< T > >
         static void val2buf( const vector< T >& val, double**buf )
         {
             double* temp = *buf;
-            *temp++ = val.size();
+            *temp++ = (double)val.size();
             for( unsigned int i = 0; i < val.size(); ++i ) {
                 Conv< T >::val2buf( val[i], &temp );
             }
